aggregate nodeapp and hostedfunction code (#88)
* aggregate nodeapp and hostedfunction code

* move deploycommon from lib/commands
